+++ core/server/master/src/main/java/alluxio/master/metastore/caching/CachingInodeStore.java
@@ -27,6 +27,7 @@ import alluxio.master.file.meta.MutableInode;
 import alluxio.master.journal.checkpoint.CheckpointInputStream;
 import alluxio.master.journal.checkpoint.CheckpointName;
 import alluxio.master.metastore.InodeStore;
+import alluxio.master.metastore.ReadOption;
 import alluxio.master.metastore.heap.HeapInodeStore;
 import alluxio.metrics.MetricsSystem;
 import alluxio.resource.LockResource;
@@ -665,7 +666,7 @@ public final class CachingInodeStore implements InodeStore, Closeable {
      * @return the ids of all children of the directory
      */
     public Collection<Long> getChildIds(Long inodeId, ReadOption option) {
-      if (!option.shouldSkipCachingAndEviction()) {
+      if (!option.shouldSkipCache()) {
         evictIfNecessary();
       }
       AtomicBoolean createdNewEntry = new AtomicBoolean(false);
@@ -698,7 +699,7 @@ public final class CachingInodeStore implements InodeStore, Closeable {
 
     private Map<String, Long> loadChildren(Long inodeId, ListingCacheEntry entry,
         ReadOption option) {
-      if (!option.shouldSkipCachingAndEviction()) {
+      if (!option.shouldSkipCache()) {
         evictIfNecessary();
       }
       entry.mModified = false;
