+++ core/server/master/src/main/java/alluxio/master/metastore/heap/HeapInodeStore.java
@@ -23,6 +23,7 @@ import alluxio.master.journal.checkpoint.CheckpointName;
 import alluxio.master.journal.checkpoint.CheckpointOutputStream;
 import alluxio.master.journal.checkpoint.CheckpointType;
 import alluxio.master.metastore.InodeStore;
+import alluxio.master.metastore.ReadOption;
 import alluxio.proto.meta.InodeMeta;
 
 import com.google.common.base.Preconditions;
